requirements.txt
Flask==3.0.3
Flask-SQLAlchemy
Werkzeug
langchain==0.3.3
langchain-huggingface==0.1.0
langchain-ollama==0.2.0
langchain-community==0.3.2
sentence-transformers==3.1.1
faiss-cpu==1.9.0
requests==2.32.3
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0
pytz==2024.2
google-generativeai
python-dotenv
transformers
torch
accelerate
peft
